moves validationSchemas and update pathways
